django_social/requirements.txt
asgiref==3.7.0
crispy-bootstrap5==0.7
Django==4.2.1
django-crispy-forms==2.0
django-environ==0.10.0
django-quill-editor==0.1.40
gunicorn==20.1.0
Pillow==9.5.0
pip==23.0.1
setuptools==66.0.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.38.4
whitenoise==6.4.0
psycopg==3.1.9